// src/types/resources/CareTeam.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TPeriod } from '../partials/Period';
import { TCareTeamParticipant } from '../partials/CareTeamParticipant';
import { TContactPoint } from '../partials/ContactPoint';
import { TAnnotation } from '../partials/Annotation';

export type TCareTeam = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status?: string;
  category?: TCodeableConcept[];
  name?: string;
  subject?: TReference;
  encounter?: TReference;
  period?: TPeriod;
  participant?: TCareTeamParticipant[];
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  managingOrganization?: TReference[];
  telecom?: TContactPoint[];
  note?: TAnnotation[];
};